Criando Listas Enviadas e controle de produtos adicionados

+++ src/screens/DetalhesLista.js
@@ -41,20 +41,28 @@ export default function ItemListScreen({ route }) {
   };
 
   const handleEditItem = (item) => {
-    setEditedItem({ ...item });
-    setIsEditModalVisible(true);
+    if (!list.sent) {
+      setEditedItem({ ...item });
+      setIsEditModalVisible(true);
+    } else {
+      Alert.alert('Erro', 'Você não pode editar itens de uma lista enviada.');
+    }
   };
 
   const handleDeleteItem = (item) => {
-    Alert.alert(
-      'Confirmar Exclusão',
-      'Tem certeza que deseja excluir este item?',
-      [
-        { text: 'Cancelar', style: 'cancel' },
-        { text: 'Excluir', style: 'destructive', onPress: () => deleteItem(item) }
-      ],
-      { cancelable: false }
-    );
+    if (!list.sent) {
+      Alert.alert(
+        'Confirmar Exclusão',
+        'Tem certeza que deseja excluir este item?',
+        [
+          { text: 'Cancelar', style: 'cancel' },
+          { text: 'Excluir', style: 'destructive', onPress: () => deleteItem(item) }
+        ],
+        { cancelable: false }
+      );
+    } else {
+      Alert.alert('Erro', 'Você não pode excluir itens de uma lista enviada.');
+    }
   };
 
   const deleteItem = (item) => {
@@ -81,6 +89,18 @@ export default function ItemListScreen({ route }) {
   };
 
   const handleEditChange = (key, value) => {
+    // Para o campo de data, mantemos a formatação enquanto permitimos apenas números
+    if (key === 'expiryDate') {
+      value = value.replace(/\D/g, '');
+      if (value.length > 2 && value.length <= 4) {
+        value = `${value.slice(0, 2)}/${value.slice(2, 4)}`;
+      } else if (value.length > 4) {
+        value = `${value.slice(0, 2)}/${value.slice(2, 4)}/${value.slice(4, 8)}`;
+      }
+    } else {
+      // Para os outros campos, permitimos apenas números
+      value = value.replace(/\D/g, '');
+    }
     setEditedItem(prevState => ({
       ...prevState,
       [key]: value
@@ -100,17 +120,19 @@ export default function ItemListScreen({ route }) {
               <Text>Data de Validade: {item.expiryDate}</Text>
               <Text>Quantidade por Embalagem: {item.quantityPerPackage}</Text>
               <Text>Quantidade: {item.quantity}</Text>
-              <View style={styles.buttonContainer}>
-                <TouchableOpacity style={styles.button} onPress={() => handleViewItem(item)}>
-                  <Text style={styles.buttonText}>Ver</Text>
-                </TouchableOpacity>
-                <TouchableOpacity style={styles.button} onPress={() => handleEditItem(item)}>
-                  <Text style={styles.buttonText}>Editar</Text>
-                </TouchableOpacity>
-                <TouchableOpacity style={styles.button} onPress={() => handleDeleteItem(item)}>
-                  <Text style={styles.buttonText}>Excluir</Text>
-                </TouchableOpacity>
-              </View>
+              {!list.sent && (
+                <View style={styles.buttonContainer}>
+                  <TouchableOpacity style={styles.button} onPress={() => handleViewItem(item)}>
+                    <Text style={styles.buttonText}>Ver</Text>
+                  </TouchableOpacity>
+                  <TouchableOpacity style={styles.button} onPress={() => handleEditItem(item)}>
+                    <Text style={styles.buttonText}>Editar</Text>
+                  </TouchableOpacity>
+                  <TouchableOpacity style={styles.button} onPress={() => handleDeleteItem(item)}>
+                    <Text style={styles.buttonText}>Excluir</Text>
+                  </TouchableOpacity>
+                </View>
+              )}
             </View>
           )}
         />
@@ -144,96 +166,100 @@ export default function ItemListScreen({ route }) {
         visible={isEditModalVisible}
         onRequestClose={closeModal}
       >
-        <View style={styles.modalContainer}>
-          <View style={styles.modalContent}>
-            <Text>Editar Item:</Text>
-            <TextInput
-              style={styles.input}
-              placeholder="Código de Barras"
-              value={editedItem?.barcode}
-              onChangeText={(text) => handleEditChange('barcode', text)}
-            />
-            <TextInput
-              style={styles.input}
-              placeholder="Data de Validade"
-              value={editedItem?.expiryDate}
-              onChangeText={(text) => handleEditChange('expiryDate', text)}
-            />
-            <TextInput
-              style={styles.input}
-              placeholder="Quantidade por Embalagem"
-              value={editedItem?.quantityPerPackage}
-              onChangeText={(text) => handleEditChange('quantityPerPackage', text)}
-            />
-            <TextInput
-              style={styles.input}
-              placeholder="Quantidade"
-              value={editedItem?.quantity}
-              onChangeText={(text) => handleEditChange('quantity', text)}
-            />
-            <Button title="Salvar" onPress={saveEditedItem} />
-            <Button title="Cancelar" onPress={closeModal} />
-          </View>
-        </View>
-      </Modal>
-    </View>
-  );
+    <View style={styles.modalContainer}>
+  <View style={styles.modalContent}>
+    <Text>Editar Item:</Text>
+    <TextInput
+      style={styles.input}
+      placeholder="Código de Barras"
+      value={editedItem?.barcode}
+      onChangeText={(text) => handleEditChange('barcode', text)}
+      keyboardType="numeric"
+    />
+    <TextInput
+      style={styles.input}
+      placeholder="Data de Validade (DD/MM/AAAA)"
+      value={editedItem?.expiryDate}
+      onChangeText={(text) => handleEditChange('expiryDate', text)}
+      keyboardType="numeric"
+    />
+    <TextInput
+      style={styles.input}
+      placeholder="Quantidade por Embalagem"
+      value={editedItem?.quantityPerPackage}
+      onChangeText={(text) => handleEditChange('quantityPerPackage', text)}
+      keyboardType="numeric"
+    />
+    <TextInput
+      style={styles.input}
+      placeholder="Quantidade"
+      value={editedItem?.quantity}
+      onChangeText={(text) => handleEditChange('quantity', text)}
+      keyboardType="numeric"
+    />
+    <Button title="Salvar" onPress={saveEditedItem} />
+    <Button title="Cancelar" onPress={closeModal} />
+  </View>
+</View>
+</Modal>
+</View>
+);
 }
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 20,
-  },
-  heading: {
-    fontSize: 20,
-    marginBottom: 20,
-  },
-  item: {
-    marginBottom: 20,
-    padding: 10,
-    borderWidth: 1,
-    borderColor: '#ccc',
-    borderRadius: 5,
-  },
-  buttonContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginTop: 10,
-  },
-  button: {
-    backgroundColor: '#007bff',
-    paddingVertical: 5,
-    paddingHorizontal: 10,
-    borderRadius: 5,
-  },
-  buttonText: {
-    color: '#fff',
-    fontSize: 16,
-  },
-  noItems: {
-    fontSize: 16,
-    color: '#999',
-  },
-  modalContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-  },
-  modalContent: {
-    backgroundColor: '#fff',
-    padding: 20,
-    borderRadius: 10,
-    elevation: 5,
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: '#ccc',
-    borderRadius: 5,
-    marginBottom: 10,
-    padding: 8,
-  },
+container: {
+  flex: 1,
+  justifyContent: 'center',
+  alignItems: 'center',
+  padding: 20,
+},
+heading: {
+  fontSize: 20,
+  marginBottom: 20,
+},
+item: {
+  marginBottom: 20,
+  padding: 10,
+  borderWidth: 1,
+  borderColor: '#ccc',
+  borderRadius: 5,
+},
+buttonContainer: {
+  flexDirection: 'row',
+  justifyContent: 'space-between',
+  marginTop: 10,
+},
+button: {
+  backgroundColor: '#007bff',
+  paddingVertical: 5,
+  paddingHorizontal: 10,
+  borderRadius: 5,
+},
+buttonText: {
+  color: '#fff',
+  fontSize: 16,
+},
+noItems: {
+  fontSize: 16,
+  color: '#999',
+},
+modalContainer: {
+  flex: 1,
+  justifyContent: 'center',
+  alignItems: 'center',
+  backgroundColor: 'rgba(0, 0, 0, 0.5)',
+},
+modalContent: {
+  backgroundColor: '#fff',
+  padding: 20,
+  borderRadius: 10,
+  elevation: 5,
+},
+input: {
+  borderWidth: 1,
+  borderColor: '#ccc',
+  borderRadius: 5,
+  marginBottom: 10,
+  padding: 8,
+},
 });
